requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
altgraph==0.17
article==0.1.1
asgiref==3.4.1
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
cryptography==36.0.2
DBUtils==2.0.2
decorator==4.4.2
Deprecated==1.2.13
Django==3.2
django-filter==21.1
django-redis==5.2.0
djangorestframework==3.12.4
et-xmlfile==1.0.1
freegames==2.3.2
frozenlist==1.3.0
future==0.18.2
h11==0.13.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
jwt==1.3.1
lxml==4.8.0
macholib==1.14
moviepy==1.0.3
multidict==6.0.2
mysqlclient==2.1.0
numpy==1.20.2
openpyxl==3.0.7
outcome==1.1.0
packaging==21.3
pefile==2021.5.13
Pillow==8.2.0
proglog==0.1.9
pycparser==2.21
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.1
PyJWT==2.3.0
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.7
pypiwin32==223
PySocks==1.7.1
pytz==2021.1
pywin32==300
pywin32-ctypes==0.2.0
redis==4.1.4
requests==2.25.1
selenium==4.1.3
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
sqlparse==0.4.2
tqdm==4.60.0
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.4
wrapt==1.13.3
wsproto==1.1.0
yarl==1.7.2
